Renamed term->golfterm
