+++ golflike/message.lua
@@ -4,9 +4,9 @@
 local utl    = require('golflike.util')
 local draw   = require('golflike.draw')
 local keymap = require('golflike.keymap')
-local termio = require('term.io')
-local keys   = require('term.keys')
-local colour = require('term.colour')
+local termio = require('golfterm.io')
+local keys   = require('golfterm.keys')
+local colour = require('golfterm.colour')
 local class  = require('30log')
 local message = class("message")
